// express-backend/src/authentication/dto/signup.dto.ts
import { z } from 'zod';

export const SignupSchema = z.object({
  firstName: z.string().min(1, 'First name is required'),
  lastName: z.string().min(1, 'Last name is required'),
  email: z.string().email('Invalid email'),
  password: z.string().min(8, 'Password must be at least 8 characters'),
});

export type SignupDto = z.infer<typeof SignupSchema>;